Correcion DBContext, Modelos y DO
Correcion DBContext, Modelos y DO

+++ Solution_BackEnd/Solution.API.W/DataModel/CategoriaProductos.cs
@@ -14,7 +14,7 @@ namespace Solution.API.W.DataModel
             Productos = new HashSet<Productos>();
         }
 
-        public decimal IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
         public string Categoria { get; set; }
 
         public virtual ICollection<Productos> Productos { get; set; }
