+++ Solution_BackEnd/Solution.API.W/DataModel/TelefonoTienda.cs
@@ -9,10 +9,10 @@ namespace Solution.API.W.DataModel
 {
     public partial class TelefonoTienda
     {
-        public decimal IdTelefono { get; set; }
+        public int IdTelefono { get; set; }
         public string Descripcion { get; set; }
         public string Numero { get; set; }
-        public decimal IdTienda { get; set; }
+        public int IdTienda { get; set; }
 
         public virtual Tiendas IdTiendaNavigation { get; set; }
     }
