+++ Solution_BackEnd/Solution.API/DataModels/CategoriaProductos.cs
@@ -12,7 +12,7 @@ namespace Solution.API.DataModels
         //    Productos = new HashSet<Productos>();
         //}
 
-        public decimal IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
         public string Categoria { get; set; }
 
         //public virtual ICollection<Productos> Productos { get; set; }
