+++ Solution_BackEnd/Solution.API/DataModels/MarcaProductos.cs
@@ -7,14 +7,16 @@ namespace Solution.API.DataModels
 {
     public class MarcaProductos
     {
-        //public MarcaProductos()
-        //{
-        //    Productos = new HashSet<Productos>();
-        //}
+       
+            ////public MarcaProductos()
+            ////{
+            ////    Productos = new HashSet<Productos>();
+            ////}
 
-        public decimal IdMarca { get; set; }
-        public string MarcaProducto { get; set; }
+            public int IdMarca { get; set; }
+            public string MarcaProducto { get; set; }
 
-        //public virtual ICollection<Productos> Productos { get; set; }
-    }
+            //public virtual ICollection<Productos> Productos { get; set; }
+        }
+    
 }
