+++ Solution_BackEnd/Solution.DO/Objects/MarcaProductos.cs
@@ -11,7 +11,7 @@ namespace Solution.DO.Objects
             Productos = new HashSet<Productos>();
         }
 
-        public decimal IdMarca { get; set; }
+        public int IdMarca { get; set; }
         public string MarcaProducto { get; set; }
 
         public virtual ICollection<Productos> Productos { get; set; }
