+++ Solution_BackEnd/Solution.DO/Objects/TelefonoTienda.cs
@@ -6,10 +6,10 @@ namespace Solution.DO.Objects
 {
     public class TelefonoTienda
     {
-        public decimal IdTelefono { get; set; }
+        public int IdTelefono { get; set; }
         public string Descripcion { get; set; }
         public string Numero { get; set; }
-        public decimal IdTienda { get; set; }
+        public int IdTienda { get; set; }
 
         public virtual Tiendas Tiendas { get; set; }
     }
