+++ Solution_BackEnd/Solution.DO/Objects/Tiendas.cs
@@ -16,7 +16,7 @@ namespace Solution.DO.Objects
             ValoracionTienda = new HashSet<ValoracionTienda>();
         }
 
-        public decimal IdTienda { get; set; }
+        public int IdTienda { get; set; }
         public string NombreTienda { get; set; }
         public string DescripcionTienda { get; set; }
