+++ Solution_BackEnd/Solution.DO/Objects/UbicacionTienda.cs
@@ -6,11 +6,11 @@ namespace Solution.DO.Objects
 {
    public class UbicacionTienda
     {
-        public decimal IdUbicacion { get; set; }
+        public int IdUbicacion { get; set; }
         public string Provincia { get; set; }
         public string Canton { get; set; }
         public string Direccion { get; set; }
-        public decimal IdTienda { get; set; }
+        public int IdTienda { get; set; }
 
         public virtual Tiendas Tiendas { get; set; }
     }
